add a new realloc test
this one is basically the same as realloc1 test,
so name it realloc 2 test. The difference is
that we do not access the offset 0, which may
make a difference.
