+++ tests/slicing-realloc3.sh
@@ -1,0 +1,6 @@
+#!/bin/bash
+
+TESTS_DIR=`dirname $0`
+source "$TESTS_DIR/test-runner.sh"
+
+run_test "sources/realloc3.c"
